There are now two functions in Activity:
reduceActivityAmout() and backgroundIncrease()

These can be overidden by subclasses. E.g. in ShopActivity, you could do:

  override def reduceActivityAmout(): Double = 10

to reduce the amount by 10 each iteration.
Or something like:

  override def reduceActivityAmout(): Double = {
    this.agent.currentActivity.   ....
    return x
  }

to calculate how much to reduce the activity by depending on the characteristics of the agent doing the activity, or whatever.

This gives us more flexibility, so later on we can configure these amounts for different agents, as well as for different particular activities.

I replaced the '+=' and '-+' functions with '++' and '--' I.e. tell the activity to increase or decrease its intensity, but the amount to change it by is defined by the backgroundIncrease() and currentIntensityDecrease() functions.

I implemented these functions in the ShopActivity to reflect the temporary changes in ABBFAgent.